[java] Moving OpenTelemetryTracer instance creation to remote.tracing.opentelemetry package to make it available to the client code too
