+++ java/client/src/org/openqa/selenium/remote/tracing/opentelemetry/BUILD.bazel
@@ -10,10 +10,14 @@ java_library(
     ],
     deps = [
         "//java/client/src/org/openqa/selenium:core",
+        "//java/client/src/org/openqa/selenium/json",
         "//java/client/src/org/openqa/selenium/remote/tracing",
         artifact("com.google.guava:guava"),
         artifact("io.opentelemetry:opentelemetry-api"),
         artifact("io.opentelemetry:opentelemetry-context"),
+        artifact("io.opentelemetry:opentelemetry-exporter-logging"),
         artifact("io.opentelemetry:opentelemetry-sdk"),
+        artifact("io.opentelemetry:opentelemetry-sdk-common"),
+        artifact("io.opentelemetry:opentelemetry-sdk-trace"),
     ],
 )
