+++ java/server/src/org/openqa/selenium/grid/log/BUILD.bazel
@@ -14,11 +14,5 @@ java_library(
         "//java/client/src/org/openqa/selenium/remote",
         "//java/server/src/org/openqa/selenium/grid/config",
         artifact("com.beust:jcommander"),
-        artifact("io.opentelemetry:opentelemetry-api"),
-        artifact("io.opentelemetry:opentelemetry-context"),
-        artifact("io.opentelemetry:opentelemetry-exporter-logging"),
-        artifact("io.opentelemetry:opentelemetry-sdk"),
-        artifact("io.opentelemetry:opentelemetry-sdk-common"),
-        artifact("io.opentelemetry:opentelemetry-sdk-trace"),
     ],
 )
