+++ java/server/src/org/openqa/selenium/grid/log/LoggingOptions.java
@@ -17,23 +17,9 @@
 
 package org.openqa.selenium.grid.log;
 
-import io.opentelemetry.api.common.Attributes;
-import io.opentelemetry.api.trace.StatusCode;
-import io.opentelemetry.api.trace.propagation.W3CTraceContextPropagator;
-import io.opentelemetry.context.propagation.ContextPropagators;
-import io.opentelemetry.sdk.OpenTelemetrySdk;
-import io.opentelemetry.sdk.common.CompletableResultCode;
-import io.opentelemetry.sdk.trace.SdkTracerProvider;
-import io.opentelemetry.sdk.trace.SpanProcessor;
-import io.opentelemetry.sdk.trace.data.EventData;
-import io.opentelemetry.sdk.trace.data.SpanData;
-import io.opentelemetry.sdk.trace.data.StatusData;
-import io.opentelemetry.sdk.trace.export.SimpleSpanProcessor;
-import io.opentelemetry.sdk.trace.export.SpanExporter;
 import org.openqa.selenium.grid.config.Config;
 import org.openqa.selenium.internal.Require;
 import org.openqa.selenium.json.Json;
-import org.openqa.selenium.json.JsonOutput;
 import org.openqa.selenium.remote.tracing.Tracer;
 import org.openqa.selenium.remote.tracing.empty.NullTracer;
 import org.openqa.selenium.remote.tracing.opentelemetry.OpenTelemetryTracer;
@@ -49,17 +35,6 @@ public class LoggingOptions {
 
   private static final Logger LOG = Logger.getLogger(LoggingOptions.class.getName());
   private static final String LOGGING_SECTION = "logging";
-
-  // We obtain the underlying tracer instance from the singleton instance
-  // that OpenTelemetry maintains. If we blindly grabbed the tracing provider
-  // and configured it, then subsequent calls would add duplicate exporters.
-  // To avoid this, stash the configured tracer on a static and weep for
-  // humanity. This implies that we're never going to need to configure
-  // tracing more than once for the entire JVM, so we're never going to be
-  // adding unit tests for this.
-  private static volatile Tracer tracer;
-
-  public static final Json JSON = new Json();
 
   private Level level = Level.INFO;
 
@@ -110,93 +85,7 @@ public class LoggingOptions {
       return new NullTracer();
     }
 
-    Tracer localTracer = tracer;
-    if (localTracer == null) {
-      synchronized (LoggingOptions.class) {
-        localTracer = tracer;
-        if (localTracer == null) {
-          localTracer = createTracer();
-          tracer = localTracer;
-        }
-      }
-    }
-    return localTracer;
-  }
-
-  private Tracer createTracer() {
-    LOG.info("Using OpenTelemetry for tracing");
-    List<SpanProcessor> exporters = new LinkedList<>();
-    exporters.add(SimpleSpanProcessor.create(new SpanExporter() {
-      @Override
-      public CompletableResultCode export(Collection<SpanData> spans) {
-
-        spans.forEach(span -> {
-          LOG.fine(String.valueOf(span));
-
-          String traceId = span.getTraceId();
-          String spanId = span.getSpanId();
-          StatusData status = span.getStatus();
-          List<EventData> eventList = span.getEvents();
-          eventList.forEach(event -> {
-            Map<String, Object> map = new HashMap<>();
-            map.put("eventTime", event.getEpochNanos());
-            map.put("traceId", traceId);
-            map.put("spanId", spanId);
-            map.put("spanKind", span.getKind().toString());
-            map.put("eventName", event.getName());
-
-            Attributes attributes = event.getAttributes();
-            map.put("attributes", attributes.asMap());
-            String jsonString = getJsonString(map);
-            if (status.getStatusCode() == StatusCode.ERROR) {
-              LOG.log(Level.WARNING, jsonString);
-            } else {
-              LOG.log(Level.FINE, jsonString);
-            }
-          });
-        });
-        return CompletableResultCode.ofSuccess();
-      }
-
-      @Override
-      public CompletableResultCode flush() {
-        return CompletableResultCode.ofSuccess();
-      }
-
-      @Override
-      public CompletableResultCode shutdown() {
-        // no-op
-        return CompletableResultCode.ofSuccess();
-      }
-    }));
-
-    // The Jaeger exporter doesn't yet have a `TracerFactoryProvider`, so we
-    // shall look up the class using reflection, and beg for forgiveness
-    // later.
-    Optional<SpanExporter> maybeJaeger = JaegerTracing.findJaegerExporter();
-    maybeJaeger.ifPresent(
-      exporter -> exporters.add(SimpleSpanProcessor.create(exporter)));
-
-    // OpenTelemetry default propagators are no-op since version 0.9.0.
-    // Hence, required propagators need to defined and added.
-    ContextPropagators propagators =
-      ContextPropagators.create((W3CTraceContextPropagator.getInstance()));
-
-    SdkTracerProvider sdkTracerProvider = SdkTracerProvider.builder()
-      .addSpanProcessor(SpanProcessor.composite(exporters))
-      .build();
-
-    OpenTelemetrySdk openTelemetrySdk = OpenTelemetrySdk.builder()
-      .setTracerProvider(sdkTracerProvider)
-      .setPropagators(propagators)
-      .buildAndRegisterGlobal();
-
-    Runtime.getRuntime()
-      .addShutdownHook(new Thread(() -> openTelemetrySdk.getTracerManagement().shutdown()));
-
-    return new OpenTelemetryTracer(
-      openTelemetrySdk.getTracer("default"),
-      propagators.getTextMapPropagator());
+    return OpenTelemetryTracer.getInstance();
   }
 
   public void configureLogging() {
@@ -265,14 +154,5 @@ public class LoggingOptions {
           }
         })
         .orElse(System.out);
-  }
-
-  private String getJsonString(Map<String, Object> map) {
-    StringBuilder text = new StringBuilder();
-    try (JsonOutput json = JSON.newOutput(text).setPrettyPrint(false)) {
-      json.write(map);
-      text.append('\n');
-    }
-    return text.toString();
   }
 }
